requirements.txt
pymupdf
openai
edge-tts
tqdm
python-dotenv
nest_asyncio
newspaper3k
lxml_html_clean
